+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project.
 
 ## [Unreleased]
 
+### Fixed
+- The `command` and `deeplink` filters now only catch messages with `UpdateType.message`.
+
 
 ## [3.4.0] - 2021-08-07